+++ Changes.md
@@ -2,6 +2,14 @@ MQ2Nav Changelog
 ================
 
 
+1.1.2
+-----
+
+**Improvements**
+* New feature: active navigation path will now be displayed on the map if MQ2Map is loaded.
+  * Color, layer and toggle added to MQ2Nav settings
+
+
 1.1.1
 -----